packages/sdk-python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tracegrade-sdk"
version = "0.1.0"
description = "TraceGrade instrumentation SDK - 3 lines to instrument your agent"
requires-python = ">=3.12"
dependencies = [
    "opentelemetry-api>=1.27",
    "opentelemetry-sdk>=1.27",
    "opentelemetry-exporter-otlp-proto-http>=1.27",
]

[project.optional-dependencies]
anthropic = ["opentelemetry-instrumentation-anthropic>=0.30"]
openai = ["opentelemetry-instrumentation-openai>=0.30"]
langchain = ["opentelemetry-instrumentation-langchain>=0.30"]
all = ["tracegrade-sdk[anthropic,openai,langchain]"]

[tool.hatch.build.targets.wheel]
packages = ["src/tracegrade_sdk"]
